вынес селекторы (надеюсь это селекторы:)) в соответствующий файл и другие правки по прошлым замечаниям
