+++ src/components/genre-list/genre-list.jsx
@@ -2,28 +2,31 @@ import React from 'react';
 import {connect} from 'react-redux';
 import PropTypes from 'prop-types';
 import {changeGenre} from '../../store/action';
-import {getGenres} from '../../utils';
+import {getGenres} from '../../store/selector';
 
 
 const GenreList = (props) => {
-  const {currentGenre, genres} = props;
+  const {currentGenre, genres, changeGenreAction} = props;
 
-  const handleGenreClick = (event) => {
+  const handleGenreClick = (genre) => {
     event.preventDefault();
-    props.dispatch(changeGenre(event.currentTarget.dataset.filter));
+    changeGenreAction(genre);
   };
 
   return (
     <ul className="catalog__genres-list">
       {genres.map((genre) => (
-        <li key={genre} className={`catalog__genres-item
-          ${currentGenre === genre
+        <li
+          key={genre}
+          className={
+            `catalog__genres-item
+            ${currentGenre === genre
           ? `catalog__genres-item--active`
           : ``}`
-        }>
+          }
+        >
           <a href="#" className="catalog__genres-link"
-            data-filter={genre}
-            onClick={handleGenreClick}
+            onClick={() => handleGenreClick(genre)}
           >
             {genre}
           </a>
@@ -34,7 +37,7 @@ const GenreList = (props) => {
 };
 
 GenreList.propTypes = {
-  dispatch: PropTypes.func.isRequired,
+  changeGenreAction: PropTypes.func.isRequired,
   currentGenre: PropTypes.string.isRequired,
   genres: PropTypes.array.isRequired,
 };
@@ -44,9 +47,9 @@ const mapStateToProps = (state) => ({
   genres: getGenres(state.movies),
 });
 
-const mapDispatchToProps = (dispatch) => ({
-  dispatch,
-});
+const mapDispatchToProps = {
+  changeGenreAction: changeGenre,
+};
 
 
 export default connect(mapStateToProps, mapDispatchToProps)(GenreList);
