+++ src/utils.js
@@ -1,6 +1,3 @@
-import {ALL_GENRES} from './const';
-
-
 export const generateId = () => Math.round(Date.now() * Math.random());
 
 export const convertMinutesToHoursAndMinutes = (totalMinutes) => {
@@ -9,25 +6,3 @@ export const convertMinutesToHoursAndMinutes = (totalMinutes) => {
 
   return `${hours}h ${minutes}m`;
 };
-
-export const getMoviesByGenre = (movies, genre) => {
-  if (genre === ALL_GENRES) {
-    return movies;
-  }
-
-  return movies.filter((movie) => movie.genre === genre);
-};
-
-export const getGenres = (movies) => {
-  const genres = movies.map((movie) => movie.genre);
-
-  return [ALL_GENRES, ...new Set(genres)];
-};
-
-export const getVisibleMovies = (movies, visibleMoviesCount) => {
-  const visibleMovies = movies.slice(0, visibleMoviesCount);
-
-  return visibleMovies;
-};
-
-export const isEnableShowMoreButton = (movies, visibleMoviesCount) => movies.length > visibleMoviesCount;
